Make search for gallery section case-insensitive
In many places, e.g. with page names and filenames, case matters
and we must stick with a case-sensitive search to avoid confusions
and mistakes.  However when searching for the gallery section
to which we have to add one or more new FPs, we can safely use
a case-insensitive search; there should never be two subheadings
on the same gallery page which differ only in case, and if this
ever happens, it would clearly be a bug on the gallery page.
Using a case-insensitive search allows us to tolerate minor spelling
mistakes which were made by the nominator, and hence to place some
more FPs correctly, without any risk of misplacing FPs [1].

Thanks to Wikimedia Commons user W.carter for proposing this change!

[1] Example of a nomination with case error in the gallery link:
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/File:Berlin_U-Bahnhof_Wittenbergplatz_Eingang-20241207-RM-155601.jpg