Handle spaces before colons in subheadings as optional
When defining the gallery link, users sometimes forget the special
space which is used before the colon in the taxonomic headings
of the biology department of our gallery pages [1].  We can easily
fine-tune the section name regex so that the bot finds the correct
gallery section regardless of that space.  Thus the bot can handle
gallery links which miss that pre-colon space (or use it in the
wrong place) without introducing a danger of false-positive matches
or other errors.

This commit also replaces the use of `wikipattern(section_name)`
in `FPCandidate.add_to_featured_list()` by a simple call of
`re.escape(section_name)`, just as we already use it in
`FPCandidate._find_gallery_insertion_place()`.  The speciality
of `wikipattern()` is that it makes the regex pattern handle
`_` and ` ` as equivalent; we do not need this here because we
clean the gallery link and replace any `_` by ` ` before calling
these methods.  So `re.escape()` is sufficient here.

[1] E.g. in this nomination:
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/File:Breeding_Double-Crested_Cormorant_Close-Up.jpg