Improve order of task handling in main()
When we run over the task arguments in `main()` to handle them,
it makes sense to handle the possible values in a meaningful order,
just as we already do some lines above during the testing of
the task arguments.  IMHO `-close` and `-park` belong together,
and `-info` fits best into the place before them because if you
ever use it you will use it before `-close` and `-park`.
So the best order is `-test`, `-info`, `-close`, `-park`.
Yes, this is a rather meaningless change, but IMHO it can help
to understand the code a tiny little bit faster and more easily.