Do not inadvertently write contents to redirect pages
In the past the FPCBot code used `page.get(get_redirect=True)`
in many places were we want to read the current contents of a page
from Wikimedia Commons.  But if the page is actually a redirect,
the argument `get_redirect=True` means that we get the wikitext
of the redirect page.  If we later modify/amend that text and write
it back to the redirect page, this means one of two things:
(i) the redirect is damaged and does not work anymore (not good);
(ii) the new things we wanted to add to the page are instead added
to the redirect page where nobody will see them, so they are lost
(not good, too).  Both errors will not be noticed by FPCBot;
errors of type (i) may be noticed by some surprised Commons user,
but errors of type (ii) will often go unnoticed for a long time
(I have found a few redirects which were inadvertently edited by
FPCBot several years ago).

For this and other reasons I have added tests for redirects in
many places where we need to read the actual contents of a page.
There are only three instances in the code which still use the
`get_redirect=True` argument; in all three cases the page should
normally be just present and not be moved, so redirects are rather
unlikely.  But exactly because of this unlikelihood we can change
these three lines of code to use `get_redirect=False` instead.
This means that if by any chance one of these page is a redirect,
Pywikibot will raise an `IsRedirectPageError` exception.

Because this will happen only very rarely or never, it seems not
really necessary to catch that exception in these three places;
so it will raise up the calling stack until it is caught by our
general exception handler which reports it on the FPC talk page.
IMHO this approach is OK here because in all three instances
it would be rather surprising and probably a sign of a severe error
if the page is actually a redirect; so this needs the attention
of experienced users, and we can get that by reporting the
exception.