Use a maximum line length of ‘90-ish’ chars
When rewriting FPCBot code, I have tried to adhere to the classic
PEP 8 rule of maximum 79 characters per line of code.  In general
this does indeed improve the code readability, and it makes it
much easier to compare different versions of the same code with
review tools that display them in parallel columns.  However,
in some cases wrapping a line in order to obey the 79 chars law
did rather impair the readability, especially when the line was
only a few chars longer than 79.

For this and other reasons various other limits have been proposed
(80, 88, 90, 100 ...).  In his well-known talk at PyCon 2015,
‘Beyond PEP 8 -- best practices for beautiful intelligible code’,
Raymond Hettinger suggests a limit of ‘90-ish characters’ as the
most useful solution [1].  With ‘90-ish’, Hettinger means a limit
around 90 characters, but I see his intentionally vague wording
also as a hint that the number should be used rather as a guiding
principle than as a fixed limit which must be applied mechanically;
the important point is not the number, but the readability.

I have experimented with loosening the limit to about 88–90 chars,
and this indeed often improves the readability and reduces the count
of lines significantly (thus making it easier to get an overview
of a method etc.).  A higher limit, like 100, clearly impairs
readability, and makes it difficult to compare different versions
of the code side by side.  Furthermore, sometimes even fully
utilizing the limit of 90 chars already seems to impair legibility,
so when a line of code is very complex or must be wrapped anyway
into several lines (e.g. the declaration of a method with several
parameters plus type hinting), it often makes sense to stay with
a lower length limit near to the traditional 79–80 chars.

Therefore I interpret Hettinger’s ‘90-ish characters’ rule
for this commit as follows:  Use the traditional 79 chars limit
as a general orientation, especially when formatting complex
method declarations, messages etc. which need to be wrapped anyway;
go up to ~ 90 chars per line as long as it does not impair the
code readability; exceed the 90 chars only in very rare cases
in which wrapping the line would *seriously* impair readability
or functionality, e.g. with paths, URLs, or `pylint: disable ...`
comments.

To utilize the handy `pycodestyle` code style checker [2] with the
new line length limit, just add a `--max-line-length=90` argument.

[1] Raymond Hettinger, ‘Beyond PEP 8 -- Best practices for beautiful
intelligible code’, PyCon 2015, at 260 s:
https://www.youtube.com/watch?v=wf-BqAjZb8M&t=260s

[2] https://pypi.org/project/pycodestyle/