Resolve redirects to moved log pages
Normally FPC log pages are not moved/renamed, so there should be
no redirects.  But we will soon introduce the splitting of the
monthly log pages if they get too long.  This requires a new
scheme for the naming of the log pages, and in order to test
the new functionality it will be necessary to run both the old
and the new code in parallel, using a redirect to allow both
versions to append nominations to the same log file despite
the different naming schemes.  To allow this the old code
must be amended by the ability to resolve (follow) redirects.