Split FPC log pages into parts to avoid overload
If a page on any Mediawiki wiki, like Wikimedia Commons, contains
a large count of templates and these templates ‘transclude’
(as this is called in the Mediawiki world) too many templates
recursively, the Mediawiki parser does not render that page
completely, but stops as soon as it has reached its threshold
for template transclusions.  This has happened from time to time
with our FPC log pages [1], and it happens more frequently now
because of the high number of FPC nominations per month.

Therefore Wikimedia Commons user W.carter has proposed [2] to split
the FPC log pages in order to avoid these rendering problems.
It seems best to let FPCBot to do the splitting, and we have agreed
on the following approach:

We define a threshold, for now 100, of max. entries per log page.
The bot starts a new log page on the 1st day of every month,
just as before, with the name ‘.../Log/<month> <year>-1’ (note
the new ‘-1’ suffix!).  If, and only if, the bot later detects
that this log page has reached 100 entries, it starts the page
‘.../Log/<month> <year>-2’ for entry no. 101 etc.  If, and only if,
it later detects that this 2nd page has again reached 100 entries,
it will create a third page for entry no. 201 etc.; and so on.

This commit implements that new approach.  It also introduces
a new header template, `{{FPC log page header}}`, which is added
by the bot at the top of every new FPC log page, explains the use
of that page and allows users to jump to the previous and next
log pages for the same month, providing a basic navigation.
The commit also makes a first step towards a more Pythonic,
exception-based error handling.  FPCBot still largely uses
a more traditional, return-value-based error handling, and the
transition to exception-based error handling will not be easy
because it requires us to rewrite large parts of the bot program.
But it seems venial to use exception-based error handling
at least for some well-defined parts of the program, and in
this commit the use of exceptions instead of various complicated
return values was a real relief and has greatly simplified
the code in the two concerning methods.

[1] E.g., and maybe for the 1st time, in
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/Log/January_2016

[2] See
https://commons.wikimedia.org/w/index.php?title=Commons_talk:Featured_picture_candidates&diff=prev&oldid=1121288942