Use a single loop for all types of nominations
Traditionally FPCBot handles all tasks (closing, parking, etc.)
in two separate loops: one loop iterates over all current delisting
nominations, a second one over all current FP nominations.
This commit unites both loops to a single one which iterates over
all desired types of nominations, regardless whether these are
only FP candidates, only delisting candidates, or both of them.
With the default setup the bot program now processes all current
nominations in a single loop; if you want to simulate the old
behavior, just start the program two times, first with the
`-delist` argument, then with the `-fpc` argument.

Advantages of this change:

1. Since June 2025 (GitHub commit 3b907f0), FPCBot processes all
nominations chronologically, i.e. in their the order of appearance
on the FP candidate list.  This has made it much simpler to follow
the bot’s operations, to spot shortcomings, and to reproduce its
behavior in tests.  However due to the separate loops for FP and
delisting nominations, the bot still processed delist nominations
before any FP nominations, even if the newest FP nomination it had
to close or park was older than the oldest delisting nomination
in the queue.  This commit removes that oddity and makes the bot
process all nominations according to their order in the list.

2. Hopefully I will soon implement full support for a third type
of nominations, delist-and-replace nominations.  (Traditionally
we handle them as a variant of delisting nominations, but in order
to achieve a solid implementation it makes more sense to introduce
a third, well-defined type of nominations.)  With the old code
I would have needed to implement a third loop just for these D&R
nominations, meaning a needless triplication of code in `main()`
and a waste of processing time.  This and some similar problems
are avoided by the changes in this commit.

3. The new code is also already prepared and optimized for the
future insertion of additional types of nominations.

4. The code in `main()` gets shorter, avoids repetitions and is
IMHO clearer than before.

Possible disadvantages:

One could fear that processing all nominations in a single loop
may consume too much memory, because we create a `Candidate` object
for every nomination which in turn links to a Pywikibot `Page`
object that may even include a cached version of the wikitext of
that page, etc.

However IMHO this danger is negligible because:

1. I have made test runs of FPCBot with candidate lists of 500
and more nominations, and never faced any problems.

2. Actually there are very few delisting nominations, they make up
less than 5% of all nominations, so handling them in a single loop
together with the FP nominations will not add much load to the
bot program.  The same will apply to the upcoming D&R nominations.

3. If we want or need to save memory, there are many other things
we can do, e.g. clearing the cached data of each nomination after
handling it.  We can also split (if this ever becomes necessary)
the loop over the nominations into several loops, each for 50
entries or so.  This would be a much clearer and more stable
approach because it would reliably prevent any overload, unlike
the traditional splitting by nomination types which would not help
at all if there would be too many concurrent FP nominations.