Clear cached page contents after use
But in case the now introduced processing of all nominations in
a single loop, together with the currently increasing frequency
of new nominations, raises concerns, we can do something simple,
which at least shouldn’t do any harm: we can clear the cached
contents of the nomination subpages.

When we create the Pywikibot `Page` objects for all nominations
in `find_candidates()` and check if some of them are redirects
etc., Pywikibot does not yet load the full contents and other
additional attributes of the pages, but just a minimum of the
information necessary for the task.  This saves memory.
Only in `check_candidates()`, when we execute the desired task
for every single nomination, we must load the wikitext and other
attributes of the pages.  We can clear the cached page contents
and other extended attributes right after the processing of each
candidate.  We also clear the instance variable `_filtered_content`
of our `Candidate` objects; we do not clear/reset the other
instance variables because that would not free much memory.

Of course we can free even more memory in an even simpler way:
in `check_candidates()`, just set each single item of the
`candidates` list to `None` after processing the `Candidate`
object in it; then Python can recycle all memory used by our
`Candidate` object and by its Pywikibot `Page` object.
I do not do this because it seems a bit irritating for my
testing and debugging, but if you are really concerned about
memory usage, here you go. ;–)