Drop separate import ‘from pywikibot import config’
In all other cases we use the full package path like in
`pywikibot.exceptions.NoPageError`, so it makes no sense
to abbreviate this one.  Furthermore, ‘config’ is a very
common term and used in many packages, so in this case
it is especially important to make clear *which* ‘config’
we are using here.