Split main() function
The `main()` function is one of the longest and most complicated
functions in the bot program, but it should be one of the
clearest ones.  We can make it more readable by splitting
it into two subroutines.  This is also necessary to prepare
for an upcoming change, namely moving our ‘last resort’
exception handler up to `main()` so that it can also report
exceptions from `find_candidates()` and its subroutines.