Move ‘last resort’ exception handler up to main()
The general ‘last resort’ exception handler in `check_candidates()`
has already proven to be useful -- its reports on the FPC talk page
have helped to discover and fix some rather obscure problems.
But it covers only exceptions which are raised during the main tasks
of the bot program.  We can allow it to catch and report also
exceptions from `find_candidates()` (which has become a much more
important complex function in the last time) if we move it up
to the `main()` function.  Now it can really cover almost any kinds
of errors.