Fix and generalize find_end_of_template()
The function `find_end_of_template()` does not only search for
the end of a template, as its name suggests, foremost it searches
for the first occurrence of (one of) the specified template(s)
and only then for the end of the template occurrence it has found.
Therefore we can make that function more useful if we tweak it
so that it returns not only the stop position after the template,
but also its start, i.e., the character index of the opening `{{`.
We need this extended functionality for one of the next commits.

Accordingly renamed the function to `find_template_pos()` and
changed its return type to either a `slice` object (containing
start and stop positions of the template) or `None`, if the template
is not found.  By adding an optional `pos` argument, the function
can also search for the next instance of the specified template(s),
allowing us to iterate over all template occurrences in a text,
or to remove all of them one by one, etc.

This commit also fixes a very stupid bug (mea culpa!): we must add
parens around `template_names` when inserting it into the search
pattern, else the search fails if `template_names` contains `|`.