Generalize strip_tag(), tidy up filter_content()
The function `strip_tag()` is very straightforward and, maybe,
a little bit too innocently trusting.  Of course normally tags
like `<s>...</s>` are used only in their most simple form within
the wikitext of discussion pages on Wikimedia Commons, but you
never know how people will use them (I have seen odd things ...).

Therefore it seems useful to make that function more flexible:
(i) by adding the `re.IGNORECASE` flag so that we can remove
also `<S>...</S>`, etc. (HTML tag names are not case-sensitive,
so the uppercase variant is plausible); and (ii) by tolerating
optional whitespace and attributes within the tags (not common
in wikitext, of course, but in accordance with the rules for
HTML tags, so it seems likely that users could use that, too,
because they remember these tags from HTML source).

The code of `filter_content()` gets a little bit clearer and
easier to understand if we group the tag removals by tag type.
Let’s also remove all comments first, because we do not need to
strip any tags etc. which are hidden inside of a comment.