Strip all unwanted HTML tags at once from the wikitext
In accordance with the new function `strip_templates()`, we can
also allow `strip_tags()` to remove all unwanted HTML tags
(incl. contents) at once from the wikitext of a nomination.

This is possible because the back reference `\1` in the regex
will match only the actual text captured by group 1 (but case-
insensitively, of course, thanks to the `re.IGNORECASE` flag).
Therefore, if we have a wikitext like
    ... <s>e.g.: <nowiki>...</nowiki></s> ...
and the `tags` parameter `"s(?:trike)?|del|nowiki"`, there is
no danger that the regex machine gets irritated by the closing
`</nowiki>` and removes only (erroneously)
    <s>e.g.: <nowiki>...</nowiki>
because group 1 captures `s` and therefore the full regex match
runs until and including `</s>`.

The new regex is not or at most marginally more complex,
the calling function `filter_content()` gets shorter and is
IHMO better structured: three lines with clear statements
to remove all comments, unwanted tags, and unwanted templates.
The new code is also faster, although this is not very important
for FPCBot which spends most of its time waiting for responses
from the Commons servers.